Allow building the code with Clang 10 on Linux
Summary:
Allow building the code with Clang 10 on Linux.

Notes:
- We are using the "uninstrumented" directory instead of "clang_uninstrumented" for Clang-only third-party dependency archives. This is because when everything is built with Clang, we should not use the "clang_" prefix for just one directory in $YB_THIRDPARTY_DIR/installed.
- Adding YB_GCC_SUFFIX and YB_CLANG_SUFFIX variables to allow appending a suffix like "-<version>" to the compiler path.

Test Plan: Jenkins

Reviewers: jharveysmith, steve.varnau

Reviewed By: steve.varnau

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D9677

+++ src/yb/gutil/strings/human_readable.cc
@@ -76,7 +76,8 @@ bool HumanReadableNumBytes::ToInt64(const string &str, int64 *num_bytes) {
       return false;
   }
   d *= scale;
-  if (d > kint64max || d < 0)
+  // We have to cast kint64max to double to avoid a warning on implicit cast that changes the value.
+  if (d > static_cast<double>(kint64max) || d < 0)
     return false;
   *num_bytes = static_cast<int64>(d + 0.5);
   if (neg) {
@@ -264,7 +265,8 @@ bool HumanReadableNum::ToDouble(const string &str, double *value) {
 bool HumanReadableInt::ToInt64(const string &str, int64 *value) {
   char *end;
   double d = strtod(str.c_str(), &end);
-  if (d > kint64max || d < kint64min)
+  // We have to cast kint64max to double to avoid a warning on implicit cast that changes the value.
+  if (d > static_cast<double>(kint64max) || d < kint64min)
     return false;
   if (*end == 'k') {
     d *= 1000;
